OC-256 Adds new WebServiceTemplateFactory and refactors existing code
