--- osgp-adapter-ws-microgrids/src/main/java/com/alliander/osgp/adapter/ws/microgrids/application/exceptions/WebServiceException.java
@@ -1,29 +1,0 @@
-/**
- * Copyright 2016 Smart Society Services B.V.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- */
-package com.alliander.osgp.adapter.ws.microgrids.application.exceptions;
-
-public class WebServiceException extends Exception {
-
-    /**
-     * Serial Version UID
-     */
-    private static final long serialVersionUID = 1L;
-
-    public WebServiceException() {
-        super();
-    }
-
-    public WebServiceException(final String message) {
-        super(message);
-    }
-
-    public WebServiceException(final String message, final Throwable t) {
-        super(message, t);
-    }
-
-}
