+++ osgp-adapter-ws-smartmetering/src/main/java/com/alliander/osgp/adapter/ws/smartmetering/infra/ws/SendNotificationServiceClient.java
@@ -14,7 +14,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import com.alliander.osgp.adapter.ws.schema.smartmetering.notification.Notification;
 import com.alliander.osgp.adapter.ws.schema.smartmetering.notification.SendNotificationRequest;
-import com.alliander.osgp.adapter.ws.smartmetering.exceptions.WebServiceSecurityException;
+import com.alliander.osgp.shared.exceptionhandling.WebServiceSecurityException;
+import com.alliander.osgp.shared.infra.ws.WebServiceTemplateFactory;
 
 public class SendNotificationServiceClient {
 
@@ -30,6 +31,7 @@ public class SendNotificationServiceClient {
      *
      * @param model
      *            The device to add.
+     * @throws WebServiceSecurityException
      * @throws GeneralSecurityException
      * @throws IOException
      */
@@ -41,6 +43,6 @@ public class SendNotificationServiceClient {
         sendNotificationRequest.setNotification(notification);
 
         this.webServiceTemplateFactory.getTemplate(organisationIdentification, notificationUsername, notificationURL)
-        .marshalSendAndReceive(sendNotificationRequest);
+                .marshalSendAndReceive(sendNotificationRequest);
     }
 }
